backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
geoalchemy2==0.14.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-multipart==0.0.5
geopandas==0.12.2
osmnx==1.3.2
aequilibrae==1.7.0
cupy-cuda11x  # Replace with the correct version for your CUDA installation or omit if CPU-only
openai==0.27.0
fastapi-jwt-auth==0.6.2
python-dotenv==1.0.0
